DEV: Added compatibility with the Glimmer Post Menu
